+++ src/js/blocks/admin/settings/forms.js
@@ -1,7 +1,13 @@
-import React, { useCallback } from "react";
-import { Form, Row, Col, Table } from "react-bootstrap";
+import React, { useState, useCallback, useMemo } from "react";
+import { Form, Row, Col, Table, ListGroup, Button } from "react-bootstrap";
 import { useFormContext, Controller, useWatch } from "react-hook-form";
 import { useSettingsContext } from "../../../app";
+import useGroupQueries from "../../../queries/groups";
+import { useWorkflowQueries } from "../../../queries/hierarchy";
+import { find, sortBy, truncate } from 'lodash';
+import { HierarchyChain } from "../../../pages/admin/hierarchy/request";
+import { Loading } from "../../components";
+
 
 export default function SettingsForms() {
     const { control } = useFormContext();
@@ -60,8 +66,91 @@ export default function SettingsForms() {
                     </Form.Group>
                 ))}
             </section>
+            <SettingsFormsDefaultRouting/>
             <SettingsFormsEmail/>
         </>
+    );
+}
+
+function SettingsFormsDefaultRouting() {
+    const [searchText,setSearchText] = useState('');
+
+    const { control, setValue } = useFormContext();
+
+    const { getGroups } = useGroupQueries();
+    const { getWorkflow } = useWorkflowQueries('form');
+    const groups = getGroups({select:d=>sortBy(d,['GROUP_NAME'])});
+    const workflows = getWorkflow({
+        enabled:!!groups.data,
+        select:d=>{
+            return d.map(w => {
+                w.GROUPS_ARRAY = w.GROUPS.split(',').map(g => {
+                    const name = find(groups.data,{GROUP_ID:g})
+                    return {GROUP_ID:g,GROUP_NAME:name?.GROUP_NAME}
+                });
+                return w;
+            });
+        },
+        initialData:[]
+    });
+
+    const listItemClick = (e,field) => {
+        e.preventDefault();
+        field.onChange(e.target.value);
+    };
+
+    const filteredWorkflows = useMemo(() => workflows.data.filter(w => w.GROUPS_ARRAY.map(g=>g.GROUP_NAME.toLowerCase()).join(' ').includes(searchText.toLocaleLowerCase())),[searchText,workflows]);
+    const selectedWorkflow = useCallback((workflowId) => workflows.data.filter(w=>w.WORKFLOW_ID==workflowId)[0],[workflows]);
+    const clearDefault = ()=>setValue('forms.defaultWorkflow','');
+    return (
+        <section>
+            <Row as="header" className="mt-3">
+                <Col as="h4">Default Routing</Col>
+            </Row>
+            {(workflows.data.length == 0) && <Loading type="alert">Loading Workflows...</Loading>}
+            {workflows.isError && <Loading type="alert" isError>Error Loading Workflows</Loading>}
+            {workflows.data.length > 0 &&
+                <>
+                    <Form.Row>
+                        <Form.Group as={Col} controlId="formWorkflowId">
+                            <Form.Label>Current Workflow: <Button title="Clear" variant="danger" style={{padding:'0.1rem 0.25rem',fontSize:'0.8rem'}} onClick={clearDefault}>X</Button></Form.Label>
+                            <Controller
+                                name="forms.defaultWorkflow"
+                                control={control}
+                                render={({field}) => (
+                                    <div className="border rounded p-3 bg-secondary-light">
+                                        <HierarchyChain list={selectedWorkflow(field.value)?.GROUPS_ARRAY} conditions={selectedWorkflow(field.value)?.CONDITIONS}/>
+                                    </div>
+                                )}
+                            />
+                        </Form.Group>
+                    </Form.Row>            
+                    <Form.Group as={Row} controlId="formWorkflowSearch">
+                        <Form.Label column md={2}>Workflow Search:</Form.Label>
+                        <Col xs="auto">
+                            <Form.Control type="search" placeholder="search workflows..." onChange={e=>setSearchText(e.target.value)} />
+                        </Col>
+                    </Form.Group>
+                    <Form.Row>
+                        <Form.Group as={Col} sm={{offset:2}} controlId="workflowListGroup">
+                        <Controller
+                            name="forms.defaultWorkflow"
+                            control={control}
+                            render={({field}) => (
+                                <ListGroup className="border list-group-condensed list-group-scrollable-25">
+                                    {filteredWorkflows.map(w =>( 
+                                        <ListGroup.Item key={w.WORKFLOW_ID} action active={field.value==w.WORKFLOW_ID} onClick={e=>listItemClick(e,field)} value={w.WORKFLOW_ID}>{w.WORKFLOW_ID}:{' '}
+                                            {truncate(w.GROUPS_ARRAY.map(g=>g.GROUP_NAME).join(' > '),{length:80,separator:' > '})}
+                                        </ListGroup.Item>))
+                                    }
+                                </ListGroup>
+                            )}
+                        />
+                        </Form.Group>
+                    </Form.Row>
+                </>
+            }
+        </section>
     );
 }
 
@@ -74,6 +163,17 @@ function SettingsFormsEmail() {
                 <Row as="header" className="mt-3">
                     <Col as="h4">Email Configuration</Col>
                 </Row>
+                <Form.Group as={Row} controlId="emailErrors">
+                    <Form.Label column md={2}>Errors Email:</Form.Label>
+                    <Col xs="auto">
+                        <Controller
+                            name='forms.email.errors'
+                            control={control}
+                            render={({field}) => <Form.Control {...field} type="email" placeholder="Enter Email Address" aria-describedby="emailErrorsHelp" />}
+                        />
+                        <Form.Text id="emailErrorsHelp" className="pt-1" muted>Email used when errors are encountered in the Forms process</Form.Text>
+                    </Col>
+                </Form.Group>
                 <Form.Group as={Row} controlId="emailEnabled">
                     <Form.Label column md={2}>Enable Notifications:</Form.Label>
                     <Col xs="auto" className="pt-2">
@@ -112,32 +212,6 @@ function SettingsFormsEmail() {
                             control={control}
                             render={({field}) => <Form.Control {...field} type="text" placeholder="Enter Email Subject" disabled={!enabled} />}
                         />
-                    </Col>
-                </Form.Group>
-                <Form.Group as={Row} controlId="emailDefault">
-                    <Form.Label column md={2}>Default Email:</Form.Label>
-                    <Col xs="auto">
-                        <Controller
-                            name='forms.email.default'
-                            control={control}
-                            render={({field}) => <Form.Control {...field} type="email" placeholder="Enter Default To Email Address" aria-describedby="emailDefaultHelp" disabled={!enabled}/>}
-                        />
-                    </Col>
-                    <Col>
-                        <Form.Text id="emailDefaultHelp" className="pt-1" muted>Default email to send to when there are no approval group members or no members set to receive email notifications.</Form.Text>
-                    </Col>
-                </Form.Group>
-                <Form.Group as={Row} controlId="emailErrors">
-                    <Form.Label column md={2}>Errors Email:</Form.Label>
-                    <Col xs="auto">
-                        <Controller
-                            name='forms.email.errors'
-                            control={control}
-                            render={({field}) => <Form.Control {...field} type="email" placeholder="Enter Email Address" aria-describedby="emailErrorsHelp" disabled={!enabled}/>}
-                        />
-                    </Col>
-                    <Col>
-                        <Form.Text id="emailErrorsHelp" className="pt-1" muted>Email to send errors to.</Form.Text>
                     </Col>
                 </Form.Group>
             </section>