+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
 ## Version 2.0.0 (DEV-202308xx)
-  * Added ability to reset Error message.
+  * Added ability to reset the Application Error message.
   * Changed the display of workflow selection for Request Hierarchy.  A scrollable list that can be filtered is now available.  The selected workflow is displayed using the graphical representation.
   * Disabled Start and End dates in user profile edit screen when editing your own profile; a user should not be able to change their own values.
-  * Added default workflow selection for Requests.  Backend code still needs to be developed to use the selection.
+  * Added default workflow settings and routing for Requests.
+  * Added default workflow settings and routing for Forms.
+  * Created config file for settings, similar to the config files for Request and Forms.  Sets default values for elements in the settings page.
 
 ## Version 2.0.0 (DEV-20230815)
   * Fixed logic error in new user caching and completed error email notification.